move validate_dataset into Dataset API, and move engine-specific logic into DatasetEngine
